Issues 96 - Moving configuration to init
