+++ mycroft/client/speech/recognizer_wrapper.py
@@ -21,7 +21,7 @@ import json
 import requests
 from speech_recognition import UnknownValueError
 
-from mycroft.configuration.config import ConfigurationManager
+from mycroft.configuration import ConfigurationManager
 from mycroft.identity import IdentityManager
 from mycroft.metrics import Stopwatch
 from mycroft.util import CerberusAccessDenied
